Create dummy map within program instead of via malloc
GNU IFUNC resolution allows for code within a program to be executed
before the entrypoint is. To allow this code to run, we must have set up
the dummy map before any execution, instead of calling malloc to do so
at the program entrypoint. This commits moves the dummy map within the
program global data.